requirements.txt
Flask==3.0.3
geopandas==0.14.3
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
plotly==5.22.0
scipy==1.13.1
psycopg2-binary
gunicorn